feat: add str trim function
